requirements.txt

gitdb==4.0.11
GitPython==3.1.43
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
streamlit==1.35.0

